# docs/openapi.json
{
  "components": {
    "schemas": {
      "AdminUsageOut": {
        "properties": {
          "users": {
            "items": {
              "$ref": "#/components/schemas/UsageOut"
            },
            "title": "Users",
            "type": "array"
          }
        },
        "required": [
          "users"
        ],
        "title": "AdminUsageOut",
        "type": "object"
      },
      "AskRequest": {
        "additionalProperties": false,
        "properties": {
          "question": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "AskRequest",
        "type": "object"
      },
      "AskResponse": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "app_version": {
            "title": "App Version",
            "type": "string"
          },
          "block_reason": {
            "anyOf": [
              {
                "enum": [
                  "injection",
                  "jailbreak",
                  "prompt_leak",
                  "pii",
                  "other"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Reason"
          },
          "blocked": {
            "title": "Blocked",
            "type": "boolean"
          },
          "bundle_version": {
            "title": "Bundle Version",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationOut"
            },
            "title": "Citations",
            "type": "array"
          },
          "citations_removed": {
            "title": "Citations Removed",
            "type": "integer"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "number"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "refused": {
            "title": "Refused",
            "type": "boolean"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "answer",
          "citations",
          "blocked",
          "block_reason",
          "refused",
          "citations_removed",
          "trace_id",
          "app_version",
          "bundle_version",
          "prompt_version",
          "model",
          "latency_ms"
        ],
        "title": "AskResponse",
        "type": "object"
      },
      "CitationOut": {
        "properties": {
          "article": {
            "title": "Article",
            "type": "string"
          },
          "excerpt": {
            "title": "Excerpt",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "paragraph": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paragraph"
          },
          "regulation": {
            "title": "Regulation",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "label",
          "regulation",
          "article",
          "paragraph",
          "excerpt",
          "url"
        ],
        "title": "CitationOut",
        "type": "object"
      },
      "DriftIndicatorsOut": {
        "properties": {
          "centroid_distance": {
            "title": "Centroid Distance",
            "type": "number"
          },
          "ks_pvalue": {
            "title": "Ks Pvalue",
            "type": "number"
          },
          "ks_statistic": {
            "title": "Ks Statistic",
            "type": "number"
          },
          "out_of_scope_ratio": {
            "title": "Out Of Scope Ratio",
            "type": "number"
          }
        },
        "required": [
          "centroid_distance",
          "out_of_scope_ratio",
          "ks_statistic",
          "ks_pvalue"
        ],
        "title": "DriftIndicatorsOut",
        "type": "object"
      },
      "DriftOut": {
        "properties": {
          "alert": {
            "title": "Alert",
            "type": "boolean"
          },
          "indicators": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DriftIndicatorsOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "ready": {
            "description": "Faux tant que la fenêtre est trop petite pour juger",
            "title": "Ready",
            "type": "boolean"
          },
          "reasons": {
            "items": {
              "enum": [
                "centroid_distance",
                "out_of_scope_ratio",
                "ks_test"
              ],
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "thresholds": {
            "$ref": "#/components/schemas/DriftThresholdsOut"
          },
          "window_size": {
            "title": "Window Size",
            "type": "integer"
          }
        },
        "required": [
          "ready",
          "window_size",
          "alert",
          "reasons",
          "indicators",
          "thresholds"
        ],
        "title": "DriftOut",
        "type": "object"
      },
      "DriftThresholdsOut": {
        "properties": {
          "centroid_distance": {
            "title": "Centroid Distance",
            "type": "number"
          },
          "ks_alpha": {
            "title": "Ks Alpha",
            "type": "number"
          },
          "min_window": {
            "title": "Min Window",
            "type": "integer"
          },
          "out_of_scope_ratio": {
            "title": "Out Of Scope Ratio",
            "type": "number"
          },
          "out_of_scope_similarity": {
            "title": "Out Of Scope Similarity",
            "type": "number"
          }
        },
        "required": [
          "centroid_distance",
          "out_of_scope_similarity",
          "out_of_scope_ratio",
          "ks_alpha",
          "min_window"
        ],
        "title": "DriftThresholdsOut",
        "type": "object"
      },
      "ErrorOut": {
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "error",
          "trace_id"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
      "FieldError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "FieldError",
        "type": "object"
      },
      "HealthOut": {
        "properties": {
          "status": {
            "const": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "HealthOut",
        "type": "object"
      },
      "ReadyOut": {
        "properties": {
          "checks": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Checks",
            "type": "object"
          },
          "status": {
            "enum": [
              "ready",
              "not_ready"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "checks"
        ],
        "title": "ReadyOut",
        "type": "object"
      },
      "UsageOut": {
        "properties": {
          "blocked": {
            "title": "Blocked",
            "type": "integer"
          },
          "cost_eur": {
            "title": "Cost Eur",
            "type": "number"
          },
          "daily_quota": {
            "title": "Daily Quota",
            "type": "integer"
          },
          "refused": {
            "title": "Refused",
            "type": "integer"
          },
          "requests": {
            "title": "Requests",
            "type": "integer"
          },
          "requests_today": {
            "title": "Requests Today",
            "type": "integer"
          },
          "tokens_in": {
            "title": "Tokens In",
            "type": "integer"
          },
          "tokens_out": {
            "title": "Tokens Out",
            "type": "integer"
          },
          "user": {
            "title": "User",
            "type": "string"
          }
        },
        "required": [
          "user",
          "requests",
          "blocked",
          "refused",
          "tokens_in",
          "tokens_out",
          "cost_eur",
          "requests_today",
          "daily_quota"
        ],
        "title": "UsageOut",
        "type": "object"
      },
      "ValidationErrorOut": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Detail",
            "type": "array"
          },
          "error": {
            "title": "Error",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "error",
          "trace_id",
          "detail"
        ],
        "title": "ValidationErrorOut",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Jeton Vigie `vig_...`",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Réponses citées sur DORA, l'AI Act, le RGPD et l'AMLR.",
    "title": "Vigie API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/healthz": {
      "get": {
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Healthz",
        "tags": [
          "ops"
        ]
      }
    },
    "/readyz": {
      "get": {
        "operationId": "readyz_readyz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadyOut"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Readyz",
        "tags": [
          "ops"
        ]
      }
    },
    "/v1/admin/drift": {
      "get": {
        "operationId": "admin_drift_v1_admin_drift_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriftOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Jeton absent, inconnu, expiré ou révoqué"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Jeton sans portée admin"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Corps de requête au-delà de 16 Ko"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOut"
                }
              }
            },
            "description": "Question vide ou trop longue"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Limite de débit ou quota quotidien atteint"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Maintenance ou modèle saturé, voir Retry-After"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Admin Drift",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/usage": {
      "get": {
        "operationId": "admin_usage_v1_admin_usage_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUsageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Jeton absent, inconnu, expiré ou révoqué"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Corps de requête au-delà de 16 Ko"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOut"
                }
              }
            },
            "description": "Question vide ou trop longue"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Limite de débit ou quota quotidien atteint"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Maintenance ou modèle saturé, voir Retry-After"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Admin Usage",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/ask": {
      "post": {
        "operationId": "ask_v1_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Jeton absent, inconnu, expiré ou révoqué"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Corps de requête au-delà de 16 Ko"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOut"
                }
              }
            },
            "description": "Question vide ou trop longue"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Limite de débit ou quota quotidien atteint"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Maintenance ou modèle saturé, voir Retry-After"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ask",
        "tags": [
          "questions"
        ]
      }
    },
    "/v1/ask/stream": {
      "post": {
        "operationId": "ask_stream_v1_ask_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Flux text/event-stream : événements delta, puis answer ou error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Jeton absent, inconnu, expiré ou révoqué"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Corps de requête au-delà de 16 Ko"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOut"
                }
              }
            },
            "description": "Question vide ou trop longue"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Limite de débit ou quota quotidien atteint"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Maintenance ou modèle saturé, voir Retry-After"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ask Stream",
        "tags": [
          "questions"
        ]
      }
    },
    "/v1/usage/me": {
      "get": {
        "operationId": "usage_me_v1_usage_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Jeton absent, inconnu, expiré ou révoqué"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Corps de requête au-delà de 16 Ko"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOut"
                }
              }
            },
            "description": "Question vide ou trop longue"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Limite de débit ou quota quotidien atteint"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Maintenance ou modèle saturé, voir Retry-After"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Usage Me",
        "tags": [
          "usage"
        ]
      }
    }
  }
}